Implement Flatbush Hilbert RTree
There are memory problems with spatial joins, potentially due to the
large number of objects created by the JTS RTree.  The JTS tree used for
the build side of spatial joins creates many objects and may not report
its memory precisely.  A Flatbush is a flattened, bulk-loaded immutable
RTree that stores the tree as a single array of doubles.  Its internal
fields are just a few primitive fields, a shared pointer to the input
geometries, a small (~5-6 size) array of ints, and a DoubleBigArray of a
size approximately the same as the input geometries.

Construction involves sorting the geometries by their Hilbert curve
index, which has good hierarchical properties and makes construction of
the rest of the tree straightforward.