Add positional filters to apply to nested columns
Add testNonNull method to TupleDomainFilter to support testing nested complex
type values for null, e.g. a[1] is null, where a is an array(array(int)).

Add NullsFilter to apply IS [NOT] NULL filter to nested complex type values.

Add PositionalFilter to apply range filter to nested columns at the bottom
level, e.g. a[1][2] > 0, where a is an array(array(int)).

Both NullsFilter and PositionalFilter are aware of the boundaries of the
top-level positions and implement getPrecedingPositionsToFail and
getSucceedingPositionsToFail methods to allow the readers to skip maximum
possible number of records at once.