Reduce excessive memory usage by ExchangeClient
ExchangeClient when it first runs will send a request for a page to
hash_partition_count nodes in the cluster. This resulted in an instance
with 10 gigabytes of data in ExchangeClients and another 8 gigabytes in
flight in jetty. This was causing full GCs since I was testing with a
40 gigabyte heap.

Previously the average response size calculation would reset to 0 every
time a 0 is sampled. This occurs quite often when a 0 response is
received from any partition that doesn't have data. This is expected to
cause ExchangeClient to poll very aggressively.

This change has ExchangeClient track the expected averageResponseSize
using an exponential moving average. This is initialized to 1 megabyte
(default page size). The alpha is defaults to 0.1.

When deciding how many nodes to contact the EMA is used and can
function as a proxy for how hard it is to find more data. The lower the
value returned by the EMA the more nodes will be contacted. This mimics
the current behavior which might not be ideal, but at least now a 0
doesn't bring the average down to 0 immediately.

The number of bytes requested is set to min(averageResponseSize * 2,
maxResponseSize). This makes the exchange client more likely to only
request a single (or fewer) pages if it ends up contacting many nodes.
Previously it would always use maxResponseSize which poorly bounded the