Use uuid for temp file when writing sorted table
Currently the temp file name for sorted table is in the format of
query_id + bucket_number + increasing_number_from_0. This has the
potential of leading to corruption when a query is running in
recoverable grouped execution mode. Assuming a task "fake died" and
we schedule the bucket to another task, then these two tasks will
be working on the same set of temp file names.

To fix this, instead of using an increasing number from 0, we use
uuid to avoid name clashing.