Check if transaction still exist on transitioning to failed query
When transaction commit fails, the transaction info is no longer in the map.

Also it fixes the race condition between calling the `transactionExists` and
the `isAutoCommit`. As the transaction may disappear in between.
The new solution is to always call `getOptionalTransactionInfo`, and then check
whether the transaction is autocommit by calling the `isAutoCommitContext` on
the returned `getOptionalTransactionInfo`.