Fix GeometryToBingTiles for certain geometries
Geometries that had a max longitude equal to `BingTile.MIN_LONGITUDE`
or min latitude equal to `BingTile.MAX_LATITUDE` would originally
not get tiled, and with 0.227 would throw an exception.  This was due to
"trimming" a SE tile to negative tileX or tileY.  This change ensures
that the SE tile cannot be west or north or the NW tile, preventing
those issues.