Use LazyBlock for columns with no filters
In cases when table scan is followed by projection with a conditional expression, it could
be beneficial to delay loading some columns as they may never be needed. For example, if
a = 2 is very selective (less than 1 in a 1000), using LazyBlock for b allows us to avoid
reading b for pages with no hits.

SELECT COUNT_IF(a = 2 AND cardinality(b) > 4 AND b[4] = 7)