Fix invalid plan for repeated lambdas in order by
Presto was creating invalid plans for queries with duplicate lambda
expressions in the order by clause because the expressions were
considered "equal" for some purposes, but not equal when converting the lambdas
to symbols.  This caused one of the expressions to not be mapped to an input
symbol. To resolve this, we don't deduplicate even seemingly equal expressions
in the order by clause so that we maintain translations for all the
expressions.

Example query:
SELECT
  COUNT(*)
FROM
  (values ARRAY['a', 'b']) as t(col1)
ORDER BY
  IF(
    SUM(
      REDUCE(
        col1,
        ROW(0),
        ("l", "r") -> "l",
        "x" -> 1
      )
    ) > 0,
    COUNT(*),
    SUM(
      REDUCE(
        col1,
        ROW(0),
        ("l", "r") -> "l",
        "x" -> 1
      )
    )
  )